Move StorageDiscoveryConfig from blivet.blivet to blivet.osinstall
As this is mostly pertaining to the installer storage configuration,
move this into osinstall.

In the interest of not breaking things between commits, I've placed a
temporary import of blivet.osinstall.StorageDiscoveryConfig in blivet.blivet,
since, for now, blivet still references it. This'll go away shortly
once those functions get moved into their rightful place.
